骨格実装クラスの整理
