+++ src/matsu/num/matrix/core/SkeletalAsymmetricOrthogonalMatrix.java
@@ -11,6 +11,9 @@
 package matsu.num.matrix.core;
 
 import java.util.Optional;
+import java.util.function.Supplier;
+
+import matsu.num.matrix.core.lazy.ImmutableLazyCacheSupplier;
 
 /**
  * 非対称な {@link OrthogonalMatrix} の骨格実装. <br>
@@ -64,10 +67,13 @@ import java.util.Optional;
  *                 この骨格実装は version 29 以降に削除される. <br>
  *                 代替となるクラスは公開されていない.
  */
+@SuppressWarnings("removal")
 @Deprecated(forRemoval = true, since = "28.6")
 public abstract class SkeletalAsymmetricOrthogonalMatrix<TT extends OrthogonalMatrix>
-        extends matsu.num.matrix.core.helper.matrix.SkeletalAsymmetricOrthogonalMatrix<TT>
         implements OrthogonalMatrix {
+
+    //逆行列(転置行列)を生成するサプライヤ
+    private final Supplier<Optional<TT>> inverseSupplier;
 
     /**
      * 唯一のコンストラクタ.
@@ -82,5 +88,70 @@ public abstract class SkeletalAsymmetricOrthogonalMatrix<TT extends OrthogonalMa
      */
     protected SkeletalAsymmetricOrthogonalMatrix() {
         super();
+        this.inverseSupplier = ImmutableLazyCacheSupplier.of(
+                () -> Optional.of(this.createTranspose()));
+        if (this instanceof Symmetric) {
+            throw new AssertionError("ImplSpec fault");
+        }
+    }
+
+    @Override
+    public final Optional<TT> inverse() {
+        return this.inverseSupplier.get();
+    }
+
+    @Override
+    public final TT transpose() {
+        return this.inverseSupplier.get().get();
+    }
+
+    /**
+     * 自身の転置行列を計算する.
+     * 
+     * <p>
+     * {@link #transpose()}, {@link #inverse()} を遅延初期化するために実装されるメソッドである. <br>
+     * それらのどちらかが初めて呼ばれたときに, 内部に持つキャッシュシステムから1度だけこのメソッドが呼ばれる. <br>
+     * 公開は禁止されており, サブクラスからもコールしてはならない.
+     *
+     * @implSpec
+     * 
+     *               {@link OrthogonalMatrix} の実装規約より,
+     * 
+     *               <blockquote>
+     * 
+     *               <pre>
+    * this.createTranspose().transpose() == this
+    * this.createTranspose().inverse().get() == this</pre>
+     * 
+     *               </blockquote>
+     * 
+     *               を満たすことが推奨される. <br>
+     *               アクセス修飾子を {@code public} にしてはいけない.
+     * 
+     * @return 自身の転置行列
+     */
+    protected abstract TT createTranspose();
+
+    /**
+     * このオブジェクトの文字列説明表現を返す.
+     * 
+     * <p>
+     * 文字列表現は明確には規定されていない(バージョン間の互換も担保されていない). <br>
+     * おそらくは次のような表現であろう. <br>
+     * {@code Matrix[dim:%dimension, orthogonal]}
+     * </p>
+     * 
+     * @implSpec
+     *               継承先においてオーバーライドを許可する. <br>
+     *               {@code Matrix["param":%param, %orthogonalType]} や
+     *               {@code Matrix["param"=%param, %orthogonalType]}
+     *               の形が適切であると思われる. <br>
+     *               {@code %orthogonalType} は "unit" などの直交行列の性質を表現する.
+     */
+    @Override
+    public String toString() {
+        return String.format(
+                "Matrix[dim: %s, orthogonal]",
+                this.matrixDimension());
     }
 }
