+++ src/matsu/num/matrix/core/SkeletalSymmetricMatrix.java
@@ -10,6 +10,8 @@
  */
 package matsu.num.matrix.core;
 
+import matsu.num.matrix.core.validation.MatrixFormatMismatchException;
+
 /**
  * {@link Symmetric} が付与された {@link Matrix} の骨格実装.
  * 
@@ -48,8 +50,8 @@ package matsu.num.matrix.core;
  *                 代替となるクラスは公開されていない.
  */
 @Deprecated(forRemoval = true, since = "28.6")
+@SuppressWarnings("removal")
 public abstract class SkeletalSymmetricMatrix<T extends SkeletalSymmetricMatrix<T>>
-        extends matsu.num.matrix.core.helper.matrix.SkeletalSymmetricMatrix<T>
         implements Matrix, Symmetric {
 
     /**
@@ -57,5 +59,52 @@ public abstract class SkeletalSymmetricMatrix<T extends SkeletalSymmetricMatrix<
      */
     protected SkeletalSymmetricMatrix() {
         super();
+    }
+
+    /**
+     * @throws MatrixFormatMismatchException {@inheritDoc}
+     * @throws NullPointerException {@inheritDoc}
+     */
+    @Override
+    public final Vector operateTranspose(Vector operand) {
+        return this.operate(operand);
+    }
+
+    /**
+     * {@code this} を返す.
+     * 
+     * <p>
+     * このメソッドの公開, サブクラスからのコールはほとんど全ての場合に不適切である.
+     * </p>
+     * 
+     * @implSpec アクセス修飾子を {@code public} にしてはいけない.
+     * 
+     * @return this
+     */
+    protected abstract T self();
+
+    @Override
+    public final T transpose() {
+        return this.self();
+    }
+
+    /**
+     * このインスタンスの文字列説明表現を返す.
+     * 
+     * <p>
+     * 文字列表現は明確には規定されていない(バージョン間の互換も担保されていない). <br>
+     * おそらくは次のような表現であろう. <br>
+     * {@code Matrix[dim:%dimension]}
+     * </p>
+     * 
+     * @implSpec
+     *               継承先においてオーバーライドを許可する. <br>
+     *               {@code Matrix["param":%param]} や
+     *               {@code Matrix["param"=%param]} の形が適切であると思われる.
+     */
+    @Override
+    public String toString() {
+        return String.format(
+                "Matrix[dim: %s]", this.matrixDimension());
     }
 }
